[API] Clipboard endpoint and resource documented
I have never written a spec before, so this is just an initial idea. If
anything is incorrect, please point me in the right direction. Also, since
I'm not necessarily working on the API, I'm not sure if I have the correct
endpoints, methods and resources. However, for this particular part, I
believe this is how we should handle it.

Also added a file for detailing common behaviour for all API requests,
and filled it with some ideas that I think will be good practice.